Mac: UI tweaks for task manager.
Nib change: Reduce row height, make scrollbars smaller.

Make text in task manager table slightly smaller.

Show in decimal digit for %cpu.

Show memory in KB/MB, not always in K.

Change update frequency from 1s to 2s to match Activity Monitor's default.

(all mac-only. ui team is fine with this.)

Finally, turn taskman on.

BUG=13156
TEST=Open task manager, look at it. Should look & feel similar to Activity Monitor.

Review URL: http://codereview.chromium.org/536038

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@36096 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/process_util_linux.cc
@@ -408,7 +408,7 @@ static int GetProcessCPU(pid_t pid) {
   return total_cpu;
 }
 
-int ProcessMetrics::GetCPUUsage() {
+double ProcessMetrics::GetCPUUsage() {
   // This queries the /proc-specific scaling factor which is
   // conceptually the system hertz.  To dump this value on another
   // system, try
